Move chosen col callback code to separate file
